+++ kmain.c
@@ -1,5 +1,5 @@
-    #include "frame_buffer.h"
-    #include "serial_port.h"
+    #include "drivers/frame_buffer.h"
+    #include "drivers/serial_port.h"
